refactor: decompose ChatClientFactory (419→178+254) into partials
Split by responsibility: public API + caching, per-provider creation methods.
